Move MIMEtype checking to its own module
